path: Fix closest point for contours
Embarrassingly, this was broken by us storing a float value
in a gsize.

Test included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8757>

+++ gsk/gskcontour.c
@@ -238,7 +238,7 @@ typedef struct
   float threshold;
   gsize idx;
   gsize best_idx;
-  gsize best_t;
+  float best_t;
 } ClosestPointData;
 
 static gboolean
